feat: search and filter refactor
