+++ src/components/search-filter-content/lodaing-state.tsx
@@ -1,0 +1,9 @@
+import { Loader2 } from "lucide-react";
+
+export default function LoadingState() {
+  return (
+    <div className="flex h-64 items-center justify-center">
+      <Loader2 className="h-8 w-8 animate-spin text-primary" />
+    </div>
+  );
+}
