+++ src/components/utils/content-utils.tsx
@@ -1,0 +1,26 @@
+/**
+ * Format date for display
+ */
+export const formatDate = (dateString: string) => {
+  if (!dateString) return "N/A";
+
+  try {
+    const date = new Date(dateString);
+    return new Intl.DateTimeFormat("en-US", {
+      year: "numeric",
+      month: "short",
+      day: "numeric",
+    }).format(date);
+  } catch (e) {
+    console.error("Error formatting date:", e);
+    return dateString;
+  }
+};
+
+/**
+ * Truncate text with ellipsis
+ */
+export const truncateText = (text: string, maxLength: number) => {
+  if (!text || text.length <= maxLength) return text;
+  return text.substring(0, maxLength) + "...";
+};
